Clarify that neat-input is integrated and link to Cabal example
I was initially confused and thought that neat-log worked well with neat-input, not that it was integrated into the module (as the input property is not documented in the readme). This minor clarification makes this clear and links to the Cabal source showing how to initialise neat-log with neat-input support.